requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
gunicorn==20.1.0
mongoengine==0.27.0
phonenumbers==8.13.7
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.1
python-environ==0.4.54
pytz==2022.7.1
sqlparse==0.2.4
whitenoise==6.4.0
